+++ contracts/token/Set.sol
@@ -5,8 +5,8 @@ pragma solidity ^0.7.6;
 library Set {
     // We define a new struct datatype that will be used to
     // hold its data in the calling contract.
-    struct Data { 
-        mapping(uint => bool) flags;
+    struct Data {
+        mapping(uint256 => bool) flags;
     }
 
     // Note that the first parameter is of type "storage
@@ -15,21 +15,19 @@ library Set {
     // special feature of library functions.  It is idiomatic
     // to call the first parameter 'self', if the function can
     // be seen as a method of that object.
-    function insert(Data storage self, uint value) public returns (bool) {
-        if (self.flags[value])
-            return false; // already there
+    function insert(Data storage self, uint256 value) public returns (bool) {
+        if (self.flags[value]) return false; // already there
         self.flags[value] = true;
         return true;
     }
 
-    function remove(Data storage self, uint value) public returns (bool) {
-        if (!self.flags[value])
-            return false; // not there
+    function remove(Data storage self, uint256 value) public returns (bool) {
+        if (!self.flags[value]) return false; // not there
         self.flags[value] = false;
         return true;
     }
 
-    function contains(Data storage self, uint value) public view returns (bool) {
+    function contains(Data storage self, uint256 value) public view returns (bool) {
         return self.flags[value];
     }
 }
